Directory reorg and modified main to initialize the classes correctly
