+++ worker/codalabworker/local_run/docker_image_manager.py
